Discover Gradle builds by looking for setting.gradle(.kts)
When the option 'gradle.nestedProjects: true' is set, the build
roots are currently discovered by looking for the wrapper scripts.
These however are completely optional for a build.

But there is one thing each Gradle build definitely needs to have:
A 'settings.gradle' or 'setting.gradle.kts' file.
Gradle itself searches for these files to accept a folder as a
Gradle build/project.

This change proposes to check for these files instead of the gradlw
scripts.

A good project to test this with is
https://github.com/microsoft/build-server-for-gradle
which contains a lot of Gradle build in the 'testProjects' folder.

Signed-off-by: Jendrik Johannes <jendrik.johannes@gmail.com>